refactor: extract media builder
